pyproject.toml
[project]
name = "aec-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-ai",
    "asyncpg",
    "sqlmodel",
    "pgvector",
    "neo4j",
    "openai",
    "tiktoken",
    "python-dotenv",
    "docling",
]
